Switched to SetFilePointerEx so we can use size_t offsets again Buffered...Stream WIP

+++ ee/io/OutputStream.h
@@ -25,10 +25,7 @@ namespace ee
 		// Returns true if this is a stream that supports seeking.
 		virtual bool CanSeek( void ) = 0;
 
-		// Note: we're using uint32_t, not size_t, for our offset and length
-		// parameters because the Windows SDK uses DWORD (ie uint32_t) values
-		// for those parameters in SetFilePointer() and WriteFile().
-		virtual bool Seek( uint32_t offset, SeekOrigin origin = SeekOrigin::kFromCurrent ) = 0;
+		virtual bool Seek( size_t offset, SeekOrigin origin = SeekOrigin::kFromCurrent ) = 0;
 
 		// Known as 'ftell' in the POSIX API
 		virtual size_t GetCurrentOffset( void ) = 0;
